pyproject.toml
[tool.poetry]
name = "workshop-airflow"
version = "0.1.0"
description = ""
authors = ["Yan Robim <yanrobim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
loguru = "^0.7.3"
streamlit = "^1.44.1"
apache-airflow = "^2.10.5"
requests = "^2.32.3"
pydantic = "^2.11.2"
python-dotenv = "^1.1.0"
apache-airflow-providers-postgres = "^6.1.1"
astro-sdk-python = {extras = ["duckdb"], version = "^1.8.1"}
duckdb = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
